add_e2e_tests
Create a static method to create and initsync an account.
